IMPORTANT: New dev environment architecture
This commit aims to simplify the use of the dev environment
and increases the velocity of compiling the Morpheo project.

Compute, Storage and Go-packages should now be present in the
parent directory, and not anymore as git submodules.

Vendoring and Building are now performed without the previous
(and great) container-based architecture, so as to gain in
compilation time.